tests/python/armadillo/pyproject.toml
[project]
name = "armadillo"
version = "1.0.0"
description = "Modern ArangoDB Testing Framework"
authors = [
    {name = "ArangoDB GmbH", email = "info@arangodb.com"},
]
license = {text = "Business Source License 1.1"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pydantic>=2.5.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "psutil>=5.9.0",
    "aiofiles>=23.0.0",
    "aiohttp>=3.9.0",
    "PyYAML>=6.0",
    "jinja2>=3.1.0",
    "pyjwt>=2.8.0",
]

[project.optional-dependencies]
monitoring = [
    "prometheus-client>=0.19.0",
    "scapy>=2.5.0",
]
analysis = [
    "pandas>=2.1.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]
development = [
    "black>=23.0.0",
    "isort>=5.13.0",
    "mypy>=1.7.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
armadillo = "armadillo.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    "arango_single: Requires single ArangoDB server",
    "arango_cluster: Requires ArangoDB cluster",
    "slow: Long-running test",
    "crash_test: Test involves crashes",
    "rta_suite: RTA test suite marker",
]

